Flashcard: Refactored ButtonView
