+++ HXD/Views/FlashcardListView.swift
@@ -8,9 +8,8 @@
 import SwiftUI
 
 struct FlashcardListView: View {
-    let TTS = TextToSpeechServices()
     @ObservedObject var flashcardVM = FlashcardViewModel()
-   
+
     var body: some View {
         VStack {
             ForEach(flashcardVM.flashcards, id: \.self) { card in
@@ -18,25 +17,13 @@ struct FlashcardListView: View {
                     Image(card)
                         .resizable()
                         .scaledToFit()
-                        .frame(width:300, height: 200)
+                        .frame(width: 300, height: 200)
                         .overlay(
-                            Button(action: {
-                                TTS.speak(text: card)
-                            }) {
-                                Circle()
-                                    .fill(Color.blue)
-                                    .frame(width: 30, height: 30)
-                                    .overlay(
-                                        Image(systemName: "speaker.wave.2.fill")
-                                            .foregroundStyle(.white)
-                                            .font(.caption)
-                                    )
-                            }
-                                .padding(.top, 150)
-                                .padding(.leading, 45)
+                            FlashcardButton(vocab: card)
+                            .padding(.top, 150)
+                            .padding(.leading, 50)
                         )
                 }
-                
             }
         }
     }
